requirements.txt
Fabric3==1.13.1.post1
Flask==0.12
Flask-MySQL==1.4.0
Flask-RESTful==0.3.5
Jinja2==2.9.5
MarkupSafe==0.23
Pillow==3.4.2
PyMySQL==0.7.10
PySocks==1.6.6
Werkzeug==0.11.15
aniso8601==1.2.0
cffi==1.9.1
click==6.7
cryptography==1.0.2
httplib2==0.10.3
idna==2.5
itsdangerous==0.24
olefile==0.44
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
pycryptodomex==3.4.5
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
six==1.10.0
twilio==5.7.0
virtualenv==15.1.0
